add “append” as synonym for “after” to type ‘a’/‘A’
That’s the word I think of for the `a` command. And the description in Vim’s `:help a` begins “*Append* text after the cursor …”, and `:help A` doesn’t include “after”, so I think “append” is the official name.

I haven’t ran the program to test this change. I just copied the format of the surrounding lines.